Player Respawn
If the player were to fall of the platform, they will now respawn. This is made by using a very simple method, a box beneath the map that is a trigger, if the player collides his position will be transformed to a specified position, which in my case is an empty gameobject in the scene.